+++ sources/engine/SiliconStudio.Paradox.Assets/ProceduralModels/ProceduralModelAsset.cs
@@ -25,9 +25,10 @@ namespace SiliconStudio.Paradox.Assets.ProceduralModels
     [ThumbnailCompiler(PreviewerCompilerNames.ProceduralModelThumbnailCompilerQualifiedName, true)]
     [AssetCompiler(typeof(ProceduralModelAssetCompiler))]
     [Display(185, "Procedural Model", "A procedural model")]
-    [AssetFormatVersion(3)]
+    [AssetFormatVersion(4)]
     [AssetUpgrader(0, 1, 2, typeof(Upgrader))]
     [AssetUpgrader(2, 3, typeof(RenameCapsuleHeight))]
+    [AssetUpgrader(3, 4, typeof(RenameDiameters))]
     public sealed class ProceduralModelAsset : Asset, IModelAsset
     {
         /// <summary>
@@ -96,8 +97,25 @@ namespace SiliconStudio.Paradox.Assets.ProceduralModels
                 var proceduralType = asset.Type;
                 if (proceduralType.Node.Tag == "!CapsuleProceduralModel" && proceduralType.Height != null)
                 {
-                    proceduralType.Length = proceduralType.Height;
+                    proceduralType.Length = 2f * (float)proceduralType.Height;
                     proceduralType.Height = DynamicYamlEmpty.Default;
+                }
+            }
+        }
+
+        class RenameDiameters : AssetUpgraderBase
+        {
+            protected override void UpgradeAsset(int currentVersion, int targetVersion, ILogger log, dynamic asset)
+            {
+                var proceduralType = asset.Type;
+                if (proceduralType.Diameter != null)
+                {
+                    proceduralType.Radius = 0.5f * (float)proceduralType.Diameter;
+                    proceduralType.Diameter = DynamicYamlEmpty.Default;
+                }
+                if (proceduralType.Node.Tag == "!TorusProceduralModel" && proceduralType.Thickness != null)
+                {
+                    proceduralType.Thickness = 0.5f * (float)proceduralType.Thickness;
                 }
             }
         }